Prototype Flask-based webapp
Attempt to stream the data back to the client to hold the connection
open while we're making requests. Not sure if this is the right approach
yet, as it allows the client to terminate the generator. Also I'm not
sure how large the buffer is. Or how to integrate this page into a
template.